backend/pyproject.toml
[tool.poetry]
name = "chore-chart-app"
version = "0.1.0"
description = "Handy dandy app to keep mah kids accountable"
authors = ["Aleksandr V Yeganov"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.30"
pydantic = "^2.7.1"
python-dotenv = "^1.0.1"
fastapi-users = {extras = ["jwt"], version = "^13.0.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
